Relocate document tests to common location
